microsvc_observe_opentelemetry/order-service/requirements.txt
flask==2.3.2
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-flask==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-exporter-otlp==1.19.0
requests==2.31.0
EOF